requirements.txt
discord.py
python-dotenv
googletrans
requests
beautifulsoup4